# dienhoatructuyen2424-be/cmd/swag/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "API DOCUMENTATION PARKING BACKEND",
        "termsOfService": "github.com/phananhtu1998/go-backend-api",
        "contact": {
            "name": "TEAM AI",
            "url": "github.com/phananhtu1998/go-backend-api",
            "email": "teamaithacoindustries@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/v1/2025",
    "paths": {
...
        "/funcpackagesmenu/create_funcpackage_menu_multiple": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Api tạo nhiều menu theo chức năng cho hệ thống",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FuncPackageMenu"
                ],
                "summary": "Tạo nhiều menu theo chức năng",
                "parameters": [
                    {
                        "description": "Funcpackage menu mapping details",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.RolesMenu"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    }
                }
            }
        },
        "/funcpackagesmenu/delete_funcpackage_menu/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Api gói chức năng menu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FuncPackageMenu"
                ],
                "summary": "Xóa chức năng menu",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID của Funcpackage menu",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    }
                }
            }
        },
        "/funcpackagesmenu/get_funcpackage_menu_by_funcpackage_id/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API này trả về chức năng menu theo ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FuncPackageMenu"
                ],
                "summary": "Lấy chức năng menu theo ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID function package",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Từ khóa tìm kiếm",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    }
                }
            }
        },
        "/funcpackagesmenu/update_funcpackage_menu/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Api cập nhật phân quyền menu cho chức năng",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FuncPackageMenu"
                ],
                "summary": "Cập nhật chức năng menu",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID của role menu",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Thông tin cần cập nhật",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RolesMenu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    }
                }
            }
        },
...
        "/upload/upload_file": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API upload file cho hệ thống",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Upload file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File cần upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    }
                }
            }
        },
        "/upload/upload_file_customer": {
            "post": {
                "security": [
                    {
                        "TokenKey": []
                    }
                ],
                "description": "API upload file cho hệ thống",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Upload file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File cần upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.RolesMenu": {
            "type": "object",
            "required": [
                "list_method",
                "menu_id",
                "role_id"
            ],
            "properties": {
                "list_method": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "['GET'",
                        "'POST']"
                    ]
                },
                "menu_id": {
                    "type": "string",
                    "example": "menu-123"
                },
                "role_id": {
                    "type": "string",
                    "example": "role-123"
                }
            }
        },
...
        "response.ErrorResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "status code",
                    "type": "integer"
                },
                "detail": {
                    "description": "dữ liệu được return"
                },
                "error": {
                    "description": "Thông báo lỗi",
                    "type": "string"
                }
            }
        },
        "response.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "status code",
                    "type": "integer"
                },
                "data": {
                    "description": "dữ liệu được return"
                },
                "message": {
                    "description": "Thông báo lỗi",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Nhập API Key vào header để xác thực API",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Nhập Bearer Token vào header. Ví dụ: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "TokenKey": {
            "description": "Nhập Token Key vào header để xác thực API",
            "type": "apiKey",
            "name": "TokenKey",
            "in": "header"
        }
    },
    "security": [
        {
            "": [
                "BearerAuth",
                "ApiKeyAuth",
                "TokenKey"
            ]
        }
    ]
}